requirements.txt
youtube-comment-downloader==0.1.76
Flask==2.1.3
Werkzeug==2.1.2
gunicorn==22.0.0
google-genai==1.16.1
